Allow to disable routing on TabbedForm
