+++ packages/ra-ui-materialui/src/form/FormTabHeader.tsx
@@ -1,10 +1,12 @@
 import * as React from 'react';
-import { useEffect } from 'react';
+import { ReactElement } from 'react';
+import PropTypes from 'prop-types';
 import { Link, useLocation } from 'react-router-dom';
 import MuiTab from '@material-ui/core/Tab';
 import classnames from 'classnames';
 import { useTranslate, useFormGroup } from 'ra-core';
-import { useForm } from 'react-final-form';
+import { useTabbedFormViewStyles } from './TabbedFormView';
+import { ClassesOverride } from '../types';
 
 export const FormTabHeader = ({
     classes,
@@ -12,26 +14,12 @@ export const FormTabHeader = ({
     value,
     icon,
     className,
+    syncWithLocation,
     ...rest
-}) => {
+}: FormTabHeaderProps): ReactElement => {
     const translate = useTranslate();
     const location = useLocation();
     const formGroup = useFormGroup(value);
-    const form = useForm();
-    const [showError, setShowError] = React.useState(false);
-
-    useEffect(() => {
-        const unsubscribe = form.subscribe(
-            state => {
-                if (!showError && (state.submitting || state.submitFailed)) {
-                    setShowError(true);
-                }
-            },
-            { submitting: true, submitFailed: true }
-        );
-
-        return unsubscribe;
-    }, [form, showError]);
 
     return (
         <MuiTab
@@ -41,14 +29,48 @@ export const FormTabHeader = ({
             className={classnames('form-tab', className, {
                 [classes.errorTabButton]:
                     formGroup.invalid &&
-                    (formGroup.touched || showError) &&
+                    formGroup.touched &&
                     location.pathname !== value,
             })}
-            component={Link}
+            component={syncWithLocation ? Link : undefined}
             to={{ ...location, pathname: value }}
             id={`tabheader-${value}`}
             aria-controls={`tabpanel-${value}`}
             {...rest}
         />
     );
+};
+
+interface FormTabHeaderProps {
+    basePath?: string;
+    className?: string;
+    classes?: ClassesOverride<typeof useTabbedFormViewStyles>;
+    hidden?: boolean;
+    icon?: ReactElement;
+    intent?: 'header' | 'content';
+    label: string;
+    margin?: 'none' | 'normal' | 'dense';
+    path?: string;
+    resource?: string;
+    syncWithLocation?: boolean;
+    value?: string;
+    variant?: 'standard' | 'outlined' | 'filled';
+}
+
+FormTabHeader.propTypes = {
+    basePath: PropTypes.string,
+    className: PropTypes.string,
+    contentClassName: PropTypes.string,
+    children: PropTypes.node,
+    intent: PropTypes.oneOf(['header', 'content']),
+    hidden: PropTypes.bool,
+    icon: PropTypes.element,
+    label: PropTypes.string.isRequired,
+    margin: PropTypes.oneOf(['none', 'dense', 'normal']),
+    path: PropTypes.string,
+    // @ts-ignore
+    record: PropTypes.object,
+    resource: PropTypes.string,
+    value: PropTypes.string,
+    variant: PropTypes.oneOf(['standard', 'outlined', 'filled']),
 };
