+++ packages/ra-ui-materialui/src/form/TabbedForm.spec.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
-import { MemoryRouter } from 'react-router-dom';
+import { MemoryRouter, Router } from 'react-router-dom';
+import { createMemoryHistory } from 'history';
 import {
     minLength,
     required,
@@ -8,8 +9,8 @@ import {
 } from 'ra-core';
 import { renderWithRedux } from 'ra-test';
 
-import TabbedForm from './TabbedForm';
-import FormTab from './FormTab';
+import { TabbedForm } from './TabbedForm';
+import { FormTab } from './FormTab';
 import TextInput from '../input/TextInput';
 import { fireEvent } from '@testing-library/react';
 
@@ -77,21 +78,23 @@ describe('<TabbedForm />', () => {
 
     it('should set the style of an inactive Tab button with errors', async () => {
         const { getAllByRole, getByLabelText } = renderWithRedux(
-            <MemoryRouter initialEntries={['/posts/12']} initialIndex={0}>
-                <TabbedForm
-                    classes={{ errorTabButton: 'error' }}
-                    resource="posts"
-                >
-                    <FormTab label="tab1">
-                        <TextInput source="title" validate={required()} />
-                    </FormTab>
-                    <FormTab label="tab2">
-                        <TextInput
-                            source="description"
-                            validate={minLength(10)}
-                        />
-                    </FormTab>
-                </TabbedForm>
+            <MemoryRouter initialEntries={['/posts/1']} initialIndex={0}>
+                <SaveContextProvider value={saveContextValue}>
+                    <TabbedForm
+                        classes={{ errorTabButton: 'error' }}
+                        resource="posts"
+                    >
+                        <FormTab label="tab1">
+                            <TextInput source="title" validate={required()} />
+                        </FormTab>
+                        <FormTab label="tab2">
+                            <TextInput
+                                source="description"
+                                validate={minLength(10)}
+                            />
+                        </FormTab>
+                    </TabbedForm>
+                </SaveContextProvider>
             </MemoryRouter>
         );
 
@@ -107,21 +110,23 @@ describe('<TabbedForm />', () => {
 
     it('should not set the style of an active Tab button with errors', () => {
         const { getAllByRole, getByLabelText } = renderWithRedux(
-            <MemoryRouter initialEntries={['/posts/12']} initialIndex={0}>
-                <TabbedForm
-                    classes={{ errorTabButton: 'error' }}
-                    resource="posts"
-                >
-                    <FormTab label="tab1">
-                        <TextInput source="title" validate={required()} />
-                    </FormTab>
-                    <FormTab label="tab2">
-                        <TextInput
-                            source="description"
-                            validate={minLength(10)}
-                        />
-                    </FormTab>
-                </TabbedForm>
+            <MemoryRouter initialEntries={['/posts/1']} initialIndex={0}>
+                <SaveContextProvider value={saveContextValue}>
+                    <TabbedForm
+                        classes={{ errorTabButton: 'error' }}
+                        resource="posts"
+                    >
+                        <FormTab label="tab1">
+                            <TextInput source="title" validate={required()} />
+                        </FormTab>
+                        <FormTab label="tab2">
+                            <TextInput
+                                source="description"
+                                validate={minLength(10)}
+                            />
+                        </FormTab>
+                    </TabbedForm>
+                </SaveContextProvider>
             </MemoryRouter>
         );
 
@@ -132,5 +137,68 @@ describe('<TabbedForm />', () => {
         fireEvent.blur(input);
         expect(tabs[0].classList.contains('error')).toEqual(false);
         expect(tabs[1].classList.contains('error')).toEqual(false);
+    });
+
+    it('should sync tabs with location by default', () => {
+        const history = createMemoryHistory({ initialEntries: ['/'] });
+
+        const { getAllByRole } = renderWithRedux(
+            <Router history={history}>
+                <SaveContextProvider value={saveContextValue}>
+                    <TabbedForm
+                        classes={{ errorTabButton: 'error' }}
+                        resource="posts"
+                    >
+                        <FormTab label="tab1">
+                            <TextInput source="title" validate={required()} />
+                        </FormTab>
+                        <FormTab label="tab2">
+                            <TextInput
+                                source="description"
+                                validate={minLength(10)}
+                            />
+                        </FormTab>
+                    </TabbedForm>
+                </SaveContextProvider>
+            </Router>
+        );
+
+        const tabs = getAllByRole('tab');
+        fireEvent.click(tabs[1]);
+        expect(history.location.pathname).toEqual('/1');
+        fireEvent.click(tabs[0]);
+        expect(history.location.pathname).toEqual('/');
+    });
+
+    it('should not sync tabs with location if syncWithLocation is false', () => {
+        const history = createMemoryHistory({ initialEntries: ['/'] });
+
+        const { getAllByRole } = renderWithRedux(
+            <Router history={history}>
+                <SaveContextProvider value={saveContextValue}>
+                    <TabbedForm
+                        classes={{ errorTabButton: 'error' }}
+                        resource="posts"
+                        syncWithLocation={false}
+                    >
+                        <FormTab label="tab1">
+                            <TextInput source="title" validate={required()} />
+                        </FormTab>
+                        <FormTab label="tab2">
+                            <TextInput
+                                source="description"
+                                validate={minLength(10)}
+                            />
+                        </FormTab>
+                    </TabbedForm>
+                </SaveContextProvider>
+            </Router>
+        );
+
+        const tabs = getAllByRole('tab');
+        fireEvent.click(tabs[1]);
+        expect(history.location.pathname).toEqual('/');
+        fireEvent.click(tabs[0]);
+        expect(history.location.pathname).toEqual('/');
     });
 });
