+++ packages/ra-ui-materialui/src/form/index.tsx
@@ -1,28 +1,23 @@
 import FormInput from './FormInput';
-import FormTab from './FormTab';
 import SimpleForm, { SimpleFormProps } from './SimpleForm';
 import SimpleFormIterator, {
     SimpleFormIteratorProps,
 } from './SimpleFormIterator';
-import TabbedForm, { TabbedFormProps } from './TabbedForm';
 import TabbedFormTabs from './TabbedFormTabs';
 import Toolbar, { ToolbarProps } from './Toolbar';
 import getFormInitialValues from './getFormInitialValues';
 export * from './FormTabHeader';
 
-export type {
-    SimpleFormProps,
-    SimpleFormIteratorProps,
-    TabbedFormProps,
-    ToolbarProps,
-};
+export * from './TabbedForm';
+export * from './FormTab';
+export * from './FormTabHeader';
+
+export type { SimpleFormProps, SimpleFormIteratorProps, ToolbarProps };
 
 export {
     FormInput,
-    FormTab,
     SimpleForm,
     SimpleFormIterator,
-    TabbedForm,
     TabbedFormTabs,
     Toolbar,
     getFormInitialValues,
